refactor : 노선 api 수정 및 통합 테스트
